Miscellaneous fixes

+++ catscripts/Government/Countries/Armenia/Armenia.sh
@@ -1,14 +1,32 @@
 #!/bin/bash
 
-egrep -i 'Armenia' newpages.txt >> Armenia.txt
+KEYWORDS_ARMENIA="Armenia"
+KEYWORDS_ARMENIA_ALL="$KEYWORDS_ARMENIA"
 
-ARMENIA=`stat --print=%s Armenia.txt`
-
-if [ $ARMENIA -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Armenia.txt"
-  export CATNAME="Armenia"
-  $CATEGORIZE
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Armenia\n"
+  fi
+
+  ARMENIA=`egrep -i "$KEYWORDS_ARMENIA" newpages.txt`
+
+  if [ "$ARMENIA" != "" ];
+  then
+    printf "$ARMENIA" > Armenia.txt
+    export CATFILE="Armenia.txt"
+    export CATNAME="Armenia"
+    $CATEGORIZE
+    rm Armenia.txt
+    unset ARMENIA
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Armenia\n"
+  fi
+
 fi
 
-rm Armenia.txt
