+++ catscripts/Government/Countries/Namibia/Namibia.sh
@@ -1,14 +1,31 @@
 #!/bin/bash
 
-egrep -i 'Namibia' newpages.txt >> Namibia.txt
+KEYWORDS_NAMIBIA="Namibia"
+KEYWORDS_NAMIBIA_ALL="$KEYWORDS_NAMIBIA"
 
-NAMIBIA=`stat --print=%s Namibia.txt`
-
-if [ $NAMIBIA -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Namibia.txt"
-  export CATNAME="Namibia"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Namibia\n"
+  fi
 
-rm Namibia.txt
+  NAMIBIA=`egrep -i "$KEYWORDS_NAMIBIA" newpages.txt`
+
+  if [ "$NAMIBIA" != "" ];
+  then
+    printf "$NAMIBIA" > Namibia.txt
+    export CATFILE="Namibia.txt"
+    export CATNAME="Namibia"
+    $CATEGORIZE
+    rm Namibia.txt
+    unset NAMIBIA
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Namibia\n"
+  fi
+
+fi
